backend/requirements.txt
fastapi
uvicorn
pyyaml
psycopg2-binary
sqlalchemy
alembic